Changed dispatch sequence to int
- to make it faster without a need to use long

+++ sd-editor/src/main/java/net/sevenscales/editor/diagram/SelectionHandler.java
@@ -58,7 +58,7 @@ public class SelectionHandler implements MouseDiagramHandler, KeyEventListener {
 	private Set<DiagramDragHandler> dragHandlers;
 	private boolean shiftOn;
 	private Diagram currentHandler;
-  private long dispachSequence;
+  private int dispachSequence;
 	private Diagram lastMultimodeSelectedDiagram;
   private boolean freehandModeOn;
   private Set<Diagram> tobeRemovedInCycle;
